Make splash screen less annoying and obnoxious
We were actively blocking the UI with sleep in Thread; this causes the app to practically hang in splash until the sleep is accomplished.

While this was okay at first, if the user were to, say, switch to another app during the splash, startActivity will still be called (as the timer is still running) and it’ll be spawned right wherever the user is. Hence, interrupting them (and surprising them, at times).

Fix this by using Runnable which allows for the app to be interrupted, but the transition to LoginActivity won’t be interrupted at all and the activity won’t be spawned if the user exits or pauses the app.

* Also add a constant delay for the animation and the runnable itself. We need them both in sync.
* Reduce the animation duration to 1000 miliseconds